+++ src/main/java/net/mcreator/radiotowers/client/gui/PanelSettingsScreen.java
@@ -292,7 +292,7 @@ public class PanelSettingsScreen extends EscScreen implements InternetStationPar
 
     @Override
     public void render(GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick) {
-        renderBackground(guiGraphics, mouseX, mouseY, partialTick);
+        renderWorldDim(guiGraphics);
         EscRect content = contentRect();
         EscPanel.renderPanel(guiGraphics, content, style);
         super.render(guiGraphics, mouseX, mouseY, partialTick);